+++ tests/algorithm_builder.rs
@@ -36,7 +36,7 @@ fn create_new_genetic_algorithm_application() {
         .with_selection(RouletteWheelSelector::new(0.7, 2))
         .with_crossover(MultiPointCrossBreeder::new(3))
         .with_mutation(RandomValueMutator::new(0.015, -2.0, 2.0))
-        .with_reinsertion(ElitistReinserter::new(MyFitnessEvaluator, false, 0.7))
+        .with_reinsertion(ElitistReinserter::new(MyFitnessEvaluator, false, 0.7).unwrap())
         .with_initial_population(initial_population)
         .build();